backend/ai_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.13.3
firebase-admin==6.3.0
python-dotenv==1.0.1
